Restricciones
restriccion de 30000, 0, repetidos y limite de 30 dias.

+++ app/src/main/java/cv/andevelopnica/com/crediveloz/DatabaseAccess.java
@@ -614,12 +614,12 @@ public class DatabaseAccess extends Activity {
 
     }
 
-    public Integer ComprbarCartera(String idCliente){
+    public Double ComprbarCartera(String idCliente){
         Cursor cursor;
         cursor = database.rawQuery("select sum(saldo) from cartera where id_cliente = "+idCliente+"",null);
         cursor.moveToFirst();
         String foundidCliente =cursor.getString(0);
-        return Integer.parseInt(foundidCliente);
+        return Double.parseDouble(foundidCliente);
 
     }
